[release] Publish a verified macOS preview with reliable recovery
Prefer the active network service, restore published configd state, reject stale process callbacks, and prove actual SNI fragmentation and isolated-service cleanup before publishing.

Tested: Prior full native arm64 and x86_64 pipeline, 17 Rust tests, 17 Swift tests, four helper scenarios, HTTPS and app launch.

Constraint: Release artifacts publish only after the exact commit passes both native architecture gates.

Not-tested: User MacBook interactions, minimum macOS 15 runtime, and Apple notarization.

+++ macos/Sources/VeilDNSCore/ProcessIdentity.swift
@@ -19,7 +19,7 @@ public struct ProcessIdentity: Codable, Equatable, Sendable {
             sysctl(buffer.baseAddress, UInt32(buffer.count), &info, &length, nil, 0)
         }
         guard pid > 1, result == 0, length == MemoryLayout<kinfo_proc>.size,
-              info.kp_proc.p_stat != SZOMB else {
+              Int32(info.kp_proc.p_stat) != SZOMB else {
             throw VeilError.message("프로세스 실행 정보를 확인할 수 없습니다.")
         }
         return .init(pid: pid, realUID: info.kp_eproc.e_pcred.p_ruid, parentPID: info.kp_eproc.e_ppid,
